moved rating stars to a component
