+++ src/components/RatingStars/styles.jsx
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const RatingStar = styled.img`
+  width: 1.5rem;
+  margin-bottom: -4px;
+`;
